refactor: move DataFiles FFI code into dedicated data_file.rs module
